support more pixel formats, refine codes, add gui_definition_unix.json
